+++ src/org/apache/mina/tcp/base/stream/SplitPackage.java
@@ -3,8 +3,8 @@ package org.apache.mina.tcp.base.stream;
 public class SplitPackage
 {
     public int    index;
-    public int    srcServerId;
-    public int    dstServerId;
+    public long   srcServerId;
+    public long   dstServerId;
     public int    messageId;
     public byte[] datas;
