+++ src/org/apache/mina/tcp/base/transserver/codec/decoder/ClientDecoderHandler.java
@@ -27,7 +27,7 @@ public class ClientDecoderHandler extends DecoderHandler
     	TCPBaseReader tcpReader        = new TCPBaseReader();
     	tcpReader.ReadHeader(reader);
     	
-    	int dstServer                  = tcpReader.GetDstServerId();
+    	long dstServer                 = tcpReader.GetDstServerId();
     	int messageId                  = tcpReader.GetMessageId();
 		
 		reader.Reset();
