+++ src/org/apache/mina/tcp/base/transserver/protocol/transmission/TServerToTServerTransReader.java
@@ -18,7 +18,7 @@ public class TServerToTServerTransReader extends TCPBaseReader
 
     
     @Override
-    public int GetSrcServerId()
+    public long GetSrcServerId()
     {
     	return TServerConfig.SERVER_ID;
     }
